fix/HC-6 워크스페이스 이동하는 기능 CommonTopMenu에서 InnerMenu로 이동
